use float16 for gunaco

+++ configs/guanaco/gunaco_llama_7B.py
@@ -29,13 +29,13 @@ model = dict(
     llm = dict(
         type=AutoModelForCausalLM.from_pretrained,
         pretrained_model_name_or_path = '/nvme/share_data/llama-7b',
-        torch_dtype = torch.float32,
+        torch_dtype = torch.float16,
         quantization_config=dict(
             type = BitsAndBytesConfig,
             load_in_4bit=True,
             load_in_8bit=False,
             llm_int8_has_fp16_weight=False,
-            bnb_4bit_compute_dtype=torch.float32,
+            bnb_4bit_compute_dtype=torch.float16,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type = 'nf4'
         )
